+++ .gitignore
@@ -6,3 +6,4 @@ dist
 .vscode-test/
 *.vsix
 .DS_Store
+deprecated
