--- packages/xbell/src/core-ioc/case.ts
@@ -1,15 +1,0 @@
-import { MetaDataType } from '../constants/index';
-import { container } from './container';
-import { PropertyKey } from '../types/index';
-
-export function Group(value: string): ClassDecorator {
-  return (target) => {
-    container.addGroup(value, target);
-  };
-}
-
-export function Case(caseName?: string): PropertyDecorator {
-  return (target, methodKey: PropertyKey) => {
-    container.addCase(caseName || methodKey, target, methodKey);
-  };
-}
