--- packages/xbell/src/core-ioc/fixture.ts
@@ -1,9 +1,0 @@
-import { genParameterDecorator } from '../utils/index';
-import { ParameterType } from '../constants/index';
-// TODO:
-export function Fixture(): MethodDecorator {
-  return (target, propertyKey) => {
-  };
-}
-
-Fixture.Param = genParameterDecorator(ParameterType.Fixture);
