--- packages/xbell/src/core-ioc/index.ts
@@ -1,10 +1,0 @@
-export * from './case';
-export * from './batch';
-export * from './depend';
-export * from './context';
-export * from './fixture';
-export * from './container';
-export * from './decorators';
-export * from './expect';
-export * from './config';
-export * from './custom';
