+++ packages/xbell/src/types/index.ts
@@ -6,4 +6,3 @@ export * from './expect';
 export * from './test';
 export * from './worker-message';
 export * from './fixture';
-export * from './project';
