--- packages/xbell/src/types/project.ts
@@ -1,3 +1,0 @@
-export type XBellProject = {
-  name: XBellProjects['names'];
-} & Omit<XBellProjects, 'names'>
